IOS-5062 Update login screen + remake navigation
