+++ Anytype/Sources/PresentationLayer/AuthNew/Login/LoginView/LoginOutput.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+@MainActor
+protocol LoginOutput: AnyObject {
+    func onOpenSettingsSelected()
+    func onOpenPublicDebugMenuSelected()
+    func onOpenQRCodeSelected(data: QrCodeScannerData)
+    func onOpenMigrationSelected(data: MigrationModuleData)
+    func onOpenSecureAlertSelected(data: SecureAlertData)
+}
